+++ database/models/initiatives.py
@@ -4,7 +4,7 @@ from decimal import Decimal
 from typing import Any
 from uuid import UUID, uuid4
 
-from sqlalchemy import ForeignKey, Integer, Numeric, String, Text, UniqueConstraint
+from sqlalchemy import BigInteger, ForeignKey, Integer, Numeric, String, Text, UniqueConstraint
 from sqlalchemy.dialects.postgresql import JSONB, UUID as PG_UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -25,7 +25,7 @@ class Initiative(Base):
     start_year: Mapped[int | None] = mapped_column("startYear", Integer, nullable=True)
     end_year: Mapped[int | None] = mapped_column("endYear", Integer, nullable=True)
     total_estimated_cost: Mapped[int | None] = mapped_column(
-        "totalEstimatedCost", Integer, nullable=True
+        "totalEstimatedCost", BigInteger, nullable=True
     )
     currency: Mapped[str | None] = mapped_column("currency", String, nullable=True)
     status: Mapped[str | None] = mapped_column("status", String, nullable=True)
